fix: cli compilation error when using cli feature
